+++ pkg/apis/deployment/v1alpha/member_status_list_test.go
@@ -66,4 +66,26 @@ func TestMemberStatusList(t *testing.T) {
 	assert.Equal(t, m1.ID, (*list)[0].ID)
 	assert.Equal(t, m2.ID, (*list)[1].ID)
 	assert.Equal(t, m3.ID, (*list)[2].ID)
+
+	list2 := &MemberStatusList{m3, m2, m1}
+	assert.True(t, list.Equal(*list2))
+	assert.True(t, list2.Equal(*list))
+
+	list3 := &MemberStatusList{m3, m1}
+	assert.False(t, list.Equal(*list3))
+	assert.False(t, list3.Equal(*list))
+
+	list4 := MemberStatusList{m3, m2, m1}
+	list4[1].Phase = "something-else"
+	assert.False(t, list.Equal(list4))
+	assert.False(t, list4.Equal(*list))
+	
+	m4 := MemberStatus{ID: "m4"}
+	list5 := &MemberStatusList{m1, m2, m4}
+	assert.False(t, list.Equal(*list5))
+	assert.False(t, list5.Equal(*list))
+
+	list6 := &MemberStatusList{m1, m2, m3, m4}
+	assert.False(t, list.Equal(*list6))
+	assert.False(t, list6.Equal(*list))
 }